SBC: LBA range check, fix some bugs in the LBA out of range check
We can not shift the LBA << 9   and compare to the file size since this mean
that for a HUGE LBA, like LBA==2^63  this will cause the 64 bit integer
to overflow and we suddenly pass all the tests and LBA sudddenly becomes LBA
Several targets have this bug as far as I can tell in testing.

Instead, use LBA as is and instead shift the filesize >> 9  before the check
to avoid this overflow.

Also, if both LBA and tranfser length are huge,  LBA + TL can wrap too
so we need to check for that too and return check condition if
lba+tl < lba

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>