spc: fix START_STOP
If power condition != 0  then we should ignore LOEJ and START bits
and not do any online/offline operations.

Fix the check for LOEJ and make sure this flag is set to 0 or 1, not 0 or 2.

Check that the removable attribute is set. We should not call
online/offline for non-removable luns

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>