MODE_SENSE: add support for mode subpages
Add support for mode subpages.
Change the modepages from an array of pointers to a linked list
so that we can store an arbitrary number od mode pages with the same
page code (but different subpage codes)

Add a simple modepage with a subpage code 0x0a/0x01 which is the
ControlExtensions subpage.
Have TCMOS set to 1 in this subpage.
This value has the following meaning (SPC):
        A timestamp changeable by methods outside this standard (TCMOS)
        bit set to one specifies that the timestamp may
        be initialized by methods outside the scope of this standard.

Which sounds appropriate since root or ntpd can change the
local time on linux without first asking TGTD and SPC for permission first

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>