COMPARE_AND_WRITE: If the compare fails we should not continue to write the data
Fix a bug in COMPARE-AND-WRITE. If the compare fails, we have to abort
with a sense code immediately and we should NOT continue to over-write
the backing store with the mismatching data.

IF the data is not matching, we also should compute at which offset
the first mismatch in the buffer occurs.  This is to be reported in
the INFORMATION field (bytes 3-7) in the sense buffer in a later
patch.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>