Fix COMPARE_AND_WRITE but leave it disabled.
Fix the compare and write opcode but leave it disabled.
It looks like consumers like vmware may be assuming that IF
compare and write is present and if it works, then the target will also
implement other opcodes like XCOPY etc.

So lets fix the command so that it works, but wait with enabling it
until we have everything else that the main consumer of this rare opcode
expects and wants.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>