ssc: make virtual tapes more closely emulate physical ones
On my physical drive (a Sony AIT-2 drive, if it matters) when
userspace reads a file mark, it gets an EOF, and the tape is advanced
past the file mark so that a subsequent read will return data from the
next file on the tape.  With a virtual drive, the file mark is
sticky--reads continue to return EOF until the tape is manually
advanced with a FSF ioctl.

If the attempt to read the next block fails, return
ASC_MEDIUM_FORMAT_CORRUPT.

Signed-off-by: Jay Fenlason <fenlason@redhat.com>
Acked-by: Mark Harvey <markh794@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>