Add support to return TimeoutDescriptors for REPORT SUPPORTED OPCODES
When an initiator requests RCTD in a REPORT SUPPORTED OPCODES call
we have to return timeout descriptors.
Previously we just returned an error if an initiator requested this
but this patch returns an empty timeout descriptor.

Since we dont really have any descriptors to return, just return an
all zero descriptor, except for the length.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>