Fix buffer_offset set to 0 in the last Data-IN carrying status
The same function was called both for setting residual counts
in data_in with status and scsi_response pdus. Bi-directional
status, being set to 0 for scsi_response when data_in is actually
passed, resets buffer_offset field inadvertently.

This precluded formatting LUNs exported thru iSCSI to a Windows
initiator.

The fix is to separate these cases and to call a specialized
function for residual count carried in data_in with status.

Signed-off-by: Alexander Nezhinsky <nezhinsky@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>