Uses size_t for open_memstream size parameter.
open_memstream expects a size_t pointer instead of an int pointer. In
64-bit platforms, size_t is 64-bit. Using the int pointer will make
other bytes be overriden when the stream is closed and on big-endian
platforms, the size will be updated to 0, which will make tgtadm create
invalid requests for tgtd.

iqn.2010-10.org.example:volume-123
tgtadm: invalid request

The size_t pointer cast hid this problem, which would produce a warning
otherwise.

Signed-off-by: Thadeu Lima de Souza Cascardo <cascardo@linux.vnet.ibm.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>