iscsi: Enable more iscsi local params
This patch is restructuring  step done in order to configuring local parameters for tgtd through tgtadm.
Local parameters in this context are parameters that are not negotiated with the initiator.

Currently there is one such parameter i.e. ISCSI_PARAM_MAX_XMIT_DLENGTH and we want to add more e.g. max commands.

This patch aims to standardize the adding and handling of such local parameters by moving them to the end of the list.
A new enumeration was named ISCSI_PARAM_FIRST_LOCAL was added.
Explicit tests done against ISCSI_PARAM_MAX_XMIT_DLENGTH in order not to send a prameter's value to the initiator
were modified to check if the parameter's index is greater or equal to ISCSI_PARAM_FIRST_LOCAL
(where the index of ISCSI_PARAM_MAX_XMIT_DLENGTH is the one of the "first local").

Signed-off-by: Shlomo Pongratz <shlomop@mellanox.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>