iscsi: Add MaxQueueCmd iscsi local parameter
Make iscsi max queue commannd configurable. This parameter is local
like "MaxXmitDataSegmentLength" i.e. it is not sent to the
initiator. The dynamic range of the parmater is [1..512] and the
default value is 128, to keep backward compatibility. The following
tgtadm command should be used in order to modify it for a certain
target

$ tgtadm --op update --mode target --tid <target-id> --name=MaxQueueCmd --value=<new-value>

Note that the new value will be applicable only to new sessions. So in
that respect, after modifying the value, if the new value need to be
used the initiator needs to logout and login again.

Signed-off-by: Shlomo Pongratz <shlomop@mellanox.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>