tgt-admin: avoid zsh-specific syntax in system() call
"lsof &> /dev/null" is apparently zsh (and maybe some other shells,
but not bash) for "redirect both stderr and stdout".  Under bash,
I believe what this means is "execute the prior command in the
background, and then execute the meaningless "> /dev/null" command".
That wouldn't be a problem except that the return code becomes the
return code of the "> /dev/null", hiding the return code of the
'lsof' command.

Fix: use the more-standard "lsof >/dev/null 2>&1" syntax.

Signed-off-by: Dan Mick <dan.mick@inktank.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>