tgtd: protect socket for management IPC with lockf(3)
If two or more tgtds are launched with same control port (a number
specified with -C option), tgtadm cannot communicate with tgtds
because the latest tgtd cominates the socket. This behavior is
confusing.

This patch lets tgtd lock the socket with lockf(3). With this change,
tgtd will exits when the socket of its control port is already locked.

Signed-off-by: Hitoshi Mitake <mitake.hitoshi@lab.ntt.co.jp>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>