The first byte in CDB USAGE DATA should be the opcode number and not 0xFF
When we build the CDB USAGE DATA field for a REPORT SUPPORTED OPCODES reply
we must set the first byte of the field to the opcode itself, and not
to 0xFF as we currently do.

SPC4:
...
The CDB USAGE DATA field contains information about the CDB for the command being queried.
The first byte of the CDB USAGE FIELD field shall contain the operation code for the command being queried.
...

A test for this is available in the libiscsi test utility :

./bin/iscsi-test-cu iscsi://127.0.0.1/iqn.ronnie.test/1 --test=SCSI.ReportSupportedOpcodes.OneCommand

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>