iscsi: Move two messages to be logged as debug and not as errors
Move two messages which are printed when connections are close from error
to debug level.

Without this change, a message currently gets logged every time an automatic
discovery process runs on the initiator node when using the discovery database.
When using multiple Control Ports, there are a lot of messages getting logged.
These essentially flood the /var/log/messages file on the initiator.

Signed-off-by: Dennis Cullison <Dennis.Cullison@Teradata.com>
Signed-off-by: Or Gerlitz <ogerlitz@mellanox.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>