SMC: add a new volume_tag parameter to storage slots
SMC barcodes in TGTD are limited to just 10 characters while T10
VolumeTags for smc storage slots allow up to 32 bytes.
Introduce a new argument when populating slots for specifying a volume tag
of up to 32 characters. If no volume tag is specified we fallback to use the
barcode as the volume tag.

Leave the current barcode as is, limited to 10 characters to not break
compatibility with scripts that might use SMC devices.

Update the dvdjukebox example script to use volume tag. This allows us to
hand a much more useful name back to the MTX tool than the barcode.
(10 characters is pretty restrictive as an ASCII text string. The full
ISO image file name is most likely much more meaningful to a human.)

Update the tgtadm manpage and show the new volume_tag argument.

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>