Validate the DataOut length in WRITE_SAME_1*
WRITE_SAME_1* takes either no data at all or exactly one block of data
as DataOut. Check for this and fail the cdb if the DataOut size differs.

(Some initiators use DataOut size == 0 (primarily together with unmap)
to mean "use a block all filled with 0" so that is why I think we
should be liberal and allow size==0 too here.)

Signed-off-by: Ronnie Sahlberg <ronniesahlberg@gmail.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>