tgtd: support pid file for tgtd daemon
Specify the pid file by command argument -p/--pid-file, tgtd can
write pid in pid file after running into daemon, and remove the
file before exiting.
The pid file's life cycle should be as the same as the tgtd, so we
can only remove it before tgtd exiting. The pid file may still remains
after tgtd crashing due to a bug, but it's does not matter. The logic
of monitor agent may be like this:
	static int tgtd_pid;

	int pid = get_pid_from_pidfile(path);
	if (!is_running(pid) || (tgdt_pid != pid)) {
		do_something();
	}

Signed-off-by: zhenwei pi <pizhenwei@bytedance.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@lab.ntt.co.jp>