tgt-admin: Fix breakage of allowing whole subnets in ACL
For example, the ACL address '192.168.1.0/24' would wrongly be
recognized as an initiator-name instead of an initiator-address by
tgt-admin, therefore not allowing access from the subnet when later
updating with tgt-admin. For CIDR notation to work we need to allow
slashes in the IP addresses.

Using slashes in ISCSI target names seems to be invalid.

Signed-off-by: Johan Frimodig <johan@frimodig.com>