fix arm64 compile breakage
This removes the hack for signalfd, which uses __NR_signalfd to check
whether a system supports signalfd or not. arm64 uses _NR_signalfd4 so
tgt wrongly assumes that arm64 doesn't support signalfd.

This hack was introduced when some distributions doesn't have a header
file for signalfd. Now all the distributions should have so we can
simply use sys/signalfd.h

Reported-by: Andreas Hasenack <andreas@canonical.com>
Signed-off-by: FUJITA Tomonori <fujita.tomonori@gmail.com>