usr: do not rely on LFS64 interfaces
LFS64 interfaces are non-standard. Relying on them causes multiple
compile errors on systems using musl libc 1.2.4 or later. They could be
worked around by defining _LARGEFILE64_SOURCE, but this only a short
term workaround, and will be removed in a future musl version.

Instead, define _FILE_OFFSET_BITS as 64, which makes all interfaces on
glibc 64-bit by default, and replace the non-standard interfaces with
the standard ones. This replaces all occurences of fstat64 with fstat,
lseek64 with lseek, pread64 with pread, pwrite64 with pwrite, and stat64
with stat.

Signed-off-by: Stijn Tintel <stijn@linux-ipv6.be>